Revert "animator: fix typing, improving docs, adding debug mode (#620)"
This reverts commit 1c1f5318b21e0b76837f54e4359b5af48d0964cc.
