+++ src/Animations/Animator.scss
@@ -41,6 +41,7 @@ $timing-micro: .12s;
 $timing-small: .15s;
 $timing-medium: .2s;
 $timing-large: .3s;
+$timing-debug: 100s;
 $timing-translate: $timing-large;
 
 @mixin timing() {
@@ -61,6 +62,10 @@ $timing-translate: $timing-large;
     transition-duration: $timing-large
   }
 
+
+  .timing-debug {
+    transition-duration: $timing-debug
+  }
 }
 
 // TimingFunction
@@ -191,12 +196,15 @@ $sequence-delay-max-child: 100;
     &.timing-large {
       opacity: .66;
     }
+    &.timing-debug {
+      opacity: 0;
+    }
   }
 }
 
 @mixin opacityEnterActive() {
   .opacity {
-    &.timing-micro, &.timing-small, &.timing-medium, &.timing-large {
+    &.timing-micro, &.timing-small, &.timing-medium, &.timing-large, &.timing-debug {
       opacity: 1;
     }
   }
@@ -227,12 +235,16 @@ $sequence-delay-max-child: 100;
       transform: scale(0.75);
     }
 
+    &.timing-debug {
+      transform: scale(0);
+    }
+
   }
 }
 
 @mixin scaleMaximum() {
   .scale {
-    &.timing-micro, &.timing-small, &.timing-medium, &.timing-large {
+    &.timing-micro, &.timing-small, &.timing-medium, &.timing-large, &.timing-debug {
       transform: scale(1);
     }
   }
@@ -243,6 +255,7 @@ $sequence-delay-max-child: 100;
 
 @mixin heightMinimum() {
   .height {
+    color: blue;
     overflow: hidden;
     max-height: 0;
   }
@@ -250,6 +263,7 @@ $sequence-delay-max-child: 100;
 
 @mixin heightMaximum() {
   .height {
+    color: red;
     overflow: hidden;
     max-height: 100%;
   }
@@ -319,6 +333,10 @@ $sequence-delay-max-child: 100;
     transition-timing-function: $easeOutCubic;
     @include translateHide();
   }
+
+  .timing-debug .translate {
+    transition-duration: 10s;
+  }
 }
 
 @mixin translateEnterActive() {
@@ -333,6 +351,10 @@ $sequence-delay-max-child: 100;
     transition-duration: .3s;
     transition-timing-function: $easeInOutQuint;
     @include translateShow();
+  }
+
+  .timing-debug .translate {
+    transition-duration: 10s;
   }
 }
