+++ src/Animations/services/helpers/child-helper.js
@@ -31,7 +31,6 @@ class ChildHelper {
 
   getLayer2() {
     return new ClassBuilder(this.data)
-      .withDebug()
       .withChild()
       .withOpacity()
       .withScale()
