+++ src/Animations/services/helpers/props-helper.js
@@ -3,7 +3,6 @@ import {
   sequences,
   directions,
   percentages,
-  debugModes,
   propsDefault,
   animationProps
 } from '../constants/constants';
@@ -36,8 +35,7 @@ const propsMap = {
   scale: scale => !!scale,
   height: height => !!height,
   translate: translate => translate && getTranslate(translate),
-  className: className => className,
-  debug: (value, name) => getDataOrDefault(debugModes, value, name)
+  className: className => className
 };
 
 class PropsHelper {
