+++ stories/Animations/Example.scss
@@ -2,9 +2,4 @@
   background: beige;
   border: 1px solid black;
   font-size: 20px;
-  padding: 5px;
-}
-
-.basic-wrapper {
-  height: 30px;
 }
