+++ stories/Animations/ShowHideControls.js
@@ -24,10 +24,6 @@ class ShowHideControls extends React.Component {
   }
 
   componentDidMount() {
-
-    if (!this.props.interval) {
-      return;
-    }
     const intervalId = setInterval(() => {
       this.toggle();
     }, 2000);
