+++ stories/Animations/index.js
@@ -4,8 +4,8 @@ import Markdown from '../utils/Components/Markdown';
 import CodeExample from '../utils/Components/CodeExample';
 import Readme from '../../src/Animations/README.md';
 
-import ExampleBasic from './ExampleBasic';
-import ExampleBasicRaw from '!raw-loader!./ExampleBasic';
+import ExampleOpacity from './ExampleOpacity';
+import ExampleOpacityRaw from '!raw-loader!./ExampleOpacity';
 import ExampleStandard from './ExampleStandard';
 import ExampleStandardRaw from '!raw-loader!./ExampleStandard';
 
@@ -15,8 +15,8 @@ storiesOf('Common', module)
       <Markdown source={Readme}/>
 
       <h1>Usage examples</h1>
-      <CodeExample title="Basic Examples" code={ExampleBasicRaw}>
-        <ExampleBasic/>
+      <CodeExample title="Basic Examples" code={ExampleOpacityRaw}>
+        <ExampleOpacity/>
       </CodeExample>
 
       <CodeExample title="Standard" code={ExampleStandardRaw}>
